% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRollUpMap.R
\name{generate.RollUpMap}
\alias{generate.RollUpMap}
\title{Generate roll up map}
\usage{
generate.RollUpMap(nProt, pep.Expr.Data)
}
\arguments{
\item{nProt}{number of proteins to map to the peptide expression data}

\item{pep.Expr.Data}{matrix of peptide expression data}
}
\value{
the peptide to protein map (for each row in 
pep.prot.Map the corresponding value corresponds 
 to the index of the protein that peptide is mapped to)
}
\description{
Tthis function generates a map for peptide to protein roll-up
}
