% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMAR.R
\name{impute.MAR}
\alias{impute.MAR}
\title{imputation under MAR/MCAR hypothesis}
\usage{
impute.MAR(dataSet.mvs, model.selector, method = "MLE")
}
\arguments{
\item{dataSet.mvs}{expression matrix containing abundances with MVs 
(either peptides or proteins)}

\item{model.selector}{binary vector; "1" indicates MAR/MCAR proteins}

\item{method}{the method to be used for MAR/MCAR missing values.
Possible values: MLE (default), SVD, KNN}
}
\value{
dataset containing only MNAR (assumed to be left-censored)
missing data
}
\description{
This function performs missing values imputation under MAR/MCAR hypothesis. 
The imputation of MVs is performed for each protein containing MAR/MCAR missing values
}
