% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impMinDet.R
\name{impute.MinDet}
\alias{impute.MinDet}
\title{Imputation with min value}
\usage{
impute.MinDet(dataSet.mvs, q = 0.01)
}
\arguments{
\item{dataSet.mvs}{expression matrix with MVs (either peptides or proteins)}

\item{q}{the q quantile used to estimate the minimum}
}
\value{
dataset containing complete abundances
}
\description{
this function performs missing values imputation by the minimum value observed
}
