% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impMinProb.R
\name{impute.MinProb}
\alias{impute.MinProb}
\title{Imputation by random draws}
\usage{
impute.MinProb(dataSet.mvs, q = 0.01, tune.sigma = 1)
}
\arguments{
\item{dataSet.mvs}{expression matrix containing abundances with 
MVs (either peptides or proteins)}

\item{q}{the q-th quantile used to estimate the minimum 
value observed for each sample}

\item{tune.sigma}{coefficient that controls the sd of the MNAR distribution}
}
\value{
dataset containing complete abundances
}
\description{
This function performs missing values imputation by random draws from a gaussian
}
