% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impQRILC.R
\name{impute.QRILC}
\alias{impute.QRILC}
\title{imputation based on quantile regression}
\usage{
impute.QRILC(dataSet.mvs, tune.sigma = 1)
}
\arguments{
\item{dataSet.mvs}{expression matrix with MVs (either peptides or proteins)}

\item{tune.sigma}{coefficient that controls the sd of the MNAR distribution}
}
\value{
a list  containing: a matrix with the 
complete abundances, a list with the 
estimated parameters of the complete
data distribution
}
\description{
this function performs missing values imputation based quantile regression
}
