% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impKNN.R
\name{impute.wrapper.KNN}
\alias{impute.wrapper.KNN}
\title{Imputation with KNN}
\usage{
impute.wrapper.KNN(dataSet.mvs, K)
}
\arguments{
\item{dataSet.mvs}{expression matrix with MVs (either peptides or proteins)}

\item{K}{the number of neighbors}
}
\value{
dataset containing complete abundances
}
\description{
This function performs missing values imputation based on KNN algorithm
}
