% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impZERO.R
\name{impute.ZERO}
\alias{impute.ZERO}
\title{Imputation by 0.}
\usage{
impute.ZERO(dataSet.mvs)
}
\arguments{
\item{dataSet.mvs}{expression matrix containing abundances with MVs 
(either peptides or proteins)}
}
\value{
dataset containing complete abundances
}
\description{
This function performs missing values imputation by 0.
}
