% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{intensity_PXD000022}
\alias{intensity_PXD000022}
\title{Dataset PXD000022 from ProteomeXchange.}
\format{
A data frame with 660 observations on the following 7 variables.
\describe{
       \item{\code{Protein.IDs}}{Peptides/Proteins names}
       \item{\code{Intensity.MB.1}}{a numeric vector}
       \item{\code{Intensity.MB.2}}{a numeric vector}
       \item{\code{Intensity.MB.3}}{a numeric vector}
       \item{\code{Intensity.MT.1}}{a numeric vector}
       \item{\code{Intensity.MT.2}}{a numeric vector}
       \item{\code{Intensity.MT.3}}{a numeric vector}
       }
}
\source{
Original MaxQuant data: 
http://proteomecentral.proteomexchange.org/cgi/GetDataset?ID=PXD000022
}
\usage{
data(intensity_PXD000022)
}
\description{
This dataset has been collected during a study designed 
to compare the protein content of the exosome-like vesicles (ELVs) released 
from C2C12 murine myoblasts during proliferation (ELV-MB), and after 
differentiation into myotuves (ELV-MT). The dataset within this package 
contains proteins intensity processed using MaxQuant. More information 
can be found on ProteomeExchange public repository 
(http://proteomecentral.proteomexchange.org/cgi/GetDataset?ID=PXD000022) 
or in the original paper (see reference).
}
\references{
Forterre A, Jalabert A, Berger E, Baudet M, Chikh K, et al. (2014) 
Proteomic Analysis of C2C12 Myoblast and Myotube Exosome-Like Vesicles: 
A New Paradigm for Myoblast-Myotube Cross Talk? PLoS ONE 9(1): 
e84153. doi:10.1371/journal.pone.0084153
}
\keyword{data}
\keyword{datasets}
