% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSelector.R
\name{model.Selector}
\alias{model.Selector}
\title{Identifies row in the data matrix affected by a MNAR missingness 
 mechanism}
\usage{
model.Selector(dataSet.mvs)
}
\arguments{
\item{dataSet.mvs}{expression matrix containing abundances with MVs 
(either peptides or proteins)}
}
\value{
flags vector; "1" denotes rows containing random missing
 values; "0" denotes rows containing left-censored missing values
}
\description{
- this function determines row in the data matrix affected by a MNAR missingness 
  mechanism
- it is based on the assumption that the distributions of the mean values  
  of proteins follows a normal distribution
- the method makes use of a decision function defined as a tradeoff between the empirical 
  CDF of the proteins' means and the theoretical CDF assuming that no MVs are present
}
