% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep2protRollUp.R
\name{pep2prot}
\alias{pep2prot}
\title{peptide to protein roll-up}
\usage{
pep2prot(pep.Expr.Data, rollup.map)
}
\arguments{
\item{pep.Expr.Data}{matrix of peptide expression data}

\item{rollup.map}{the map to peptide to protein mapping}
}
\value{
matrix of peptide expression data
}
\description{
this function performs peptide to protein roll-up
}
