% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePepData.R
\name{generate.ExpressionData}
\alias{generate.ExpressionData}
\title{Generate expression data}
\usage{
generate.ExpressionData(
  nSamples1,
  nSamples2,
  meanSamples,
  sdSamples,
  nFeatures,
  nFeaturesUp,
  nFeaturesDown,
  meanDynRange,
  sdDynRange,
  meanDiffAbund,
  sdDiffAbund
)
}
\arguments{
\item{nSamples1}{number of samples in condition 1}

\item{nSamples2}{number of samples in condition 2}

\item{meanSamples}{xxx}

\item{sdSamples}{xxx}

\item{nFeatures}{number of total features}

\item{nFeaturesUp}{number of features up regulated}

\item{nFeaturesDown}{number of features down regulated}

\item{meanDynRange}{mean value of the dynamic range}

\item{sdDynRange}{sd of the dynamic range}

\item{meanDiffAbund}{xxx}

\item{sdDiffAbund}{xxx}
}
\value{
A list containing the data, the conditions label and the regulation
label (up/down/no)
}
\description{
this function generates artificial peptide abundance data with DA proteins
samples are drawn from a gaussian distribution
}
