% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMAR-MNAR.R
\name{impute.MAR.MNAR}
\alias{impute.MAR.MNAR}
\title{Imputation under MCAR and MNAR hypothesis}
\usage{
impute.MAR.MNAR(
  dataSet.mvs,
  model.selector,
  method.MAR = "KNN",
  method.MNAR = "QRILC"
)
}
\arguments{
\item{dataSet.mvs}{expression matrix containing abundances with MVs 
(either peptides or proteins)}

\item{model.selector}{- binary vector; "1" indicates MCAR proteins}

\item{method.MAR}{- the method to be used for MAR missing values
- possible values: MLE (default), SVD, KNN}

\item{method.MNAR}{- the method to be used for MAR missing values}
}
\value{
dataset containing complete abundances
}
\description{
this function performs missing values imputation under MCAR and MNAR hypothesis
}
