% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impMLE.R
\name{impute.wrapper.MLE}
\alias{impute.wrapper.MLE}
\title{imputation using the EM algorithm}
\usage{
impute.wrapper.MLE(dataSet.mvs)
}
\arguments{
\item{dataSet.mvs}{expression matrix with MVs (either peptides or proteins)}
}
\value{
expression matrix with MVs imputed
}
\description{
This function performs missing values imputation using the EM algorithm
}
