% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impSVD.R
\name{impute.wrapper.SVD}
\alias{impute.wrapper.SVD}
\title{imputation based on SVD algorithm}
\usage{
impute.wrapper.SVD(dataSet.mvs, K)
}
\arguments{
\item{dataSet.mvs}{expression matrix with MVs (either peptides or proteins)}

\item{K}{the number of PCs}
}
\value{
expression matrix with MVs imputed
}
\description{
this function performs missing values imputation based on SVD algorithm
}
